feat: Integrar gestión de departamentos en la página de Departamentos y eliminar componentes obsoletos
